requirements.txt
alpaca-trade-api>=3.0.0
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
colorama>=0.4.6
matplotlib>=3.7.0
